Update README.md
This is a very simple and fun memory game!
The player has 3 seconds to press the buttons to match the red LED on the left side and then press the "Final Answear" button to submit their choice. If it is correct, the answear will be counted and displayed on the right side of the page. If not, be faster next time! :)